+++ src/Marvin.Migrations.PostgreSQL/PostgreDbProvider.cs
@@ -89,6 +89,10 @@ namespace Marvin.Migrations.PostgreSQL
             {
                 throw new MigrationException(MigrationError.ConnectionError, $"Can not connect to DB {DbName}", e);
             }
+            catch (InvalidOperationException)
+            {
+                throw;
+            }
             catch (Exception e)
             {
                 throw new MigrationException(MigrationError.CreatingDbError, $"Can not create database {DbName}", e);
@@ -101,13 +105,14 @@ namespace Marvin.Migrations.PostgreSQL
         {
             try
             {
-                var result = await InternalExecuteScalarScriptAsync(_connection, String.Format(CheckDbExistQueryFormat, DbName))
+                var result = await ExecuteScalarScriptWithoutInitialCatalogAsync(String.Format(CheckDbExistQueryFormat, DbName))
                     .ConfigureAwait(false);
                 if (result == null || (Int32) result != 1)
                 {
                     return DbState.NotCreated;
                 }
 
+                AssertConnection(_connection);
                 var dbVersion = await InternalGetDbVersionAsync()
                     .ConfigureAwait(false);
                 if (dbVersion == null) return DbState.Outdated;
@@ -125,8 +130,6 @@ namespace Marvin.Migrations.PostgreSQL
         /// <inheritdoc />
         public async Task CreateDatabaseIfNotExistsAsync()
         {
-            AssertConnection();
-
             var createDbQueryFormat = "CREATE DATABASE \"{0}\" "
                                       + @"WITH 
                            ENCODING = '{1}'
@@ -143,10 +146,10 @@ namespace Marvin.Migrations.PostgreSQL
             try
             {
                 var result =
-                    await InternalExecuteScalarScriptAsync(_connection, String.Format(CheckDbExistQueryFormat, DbName));
-                if (result == null || (Int32) result != 1)
+                    await ExecuteScalarScriptWithoutInitialCatalogAsync(String.Format(CheckDbExistQueryFormat, DbName));
+                if (result == null || result is int i && i != 1 || result is bool b && !b)
                 {
-                    await InternalExecuteScriptAsync(createDbQueryString);
+                    await ExecuteScriptWithoutInitialCatalogAsync(createDbQueryString);
                 }
             }
             catch (PostgresException e)
@@ -168,6 +171,10 @@ namespace Marvin.Migrations.PostgreSQL
             {
                 throw new MigrationException(MigrationError.ConnectionError, $"Can not connect to DB {DbName}", e);
             }
+            catch (InvalidOperationException)
+            {
+                throw;
+            }
             catch (Exception e)
             {
                 throw new MigrationException(MigrationError.CreatingDbError, $"Can not create database {DbName}", e);
@@ -180,8 +187,8 @@ namespace Marvin.Migrations.PostgreSQL
             try
             {
                 var result =
-                    await InternalExecuteScalarScriptAsync(_connection, String.Format(CheckDbExistQueryFormat, DbName));
-                return result != null && (Int32) result == 1;
+                    await ExecuteScalarScriptWithoutInitialCatalogAsync(String.Format(CheckDbExistQueryFormat, DbName));
+                return result != null && (result is int i && i == 1 || result is bool b && b);
             }
             catch (PostgresException e)
                 when (e.SqlState.StartsWith("08")
@@ -198,15 +205,19 @@ namespace Marvin.Migrations.PostgreSQL
                 throw new MigrationException(MigrationError.AuthorizationError,
                     $"Invalid authorization specification for {DbName}", e);
             }
+            catch (InvalidOperationException)
+            {
+                throw;
+            }
             catch (Exception e)
             {
                 throw new MigrationException(MigrationError.Unknown, $"Can not check existence of {DbName} database", e);
             }
         }
 
-        private void AssertConnection()
+        private void AssertConnection(NpgsqlConnection connection)
         {
-            if (_connection == null || _connection.State == ConnectionState.Closed || _connection.State == ConnectionState.Broken)
+            if (connection == null || connection.State == ConnectionState.Closed || connection.State == ConnectionState.Broken)
                 throw new InvalidOperationException($"Connection is not opened. Use {nameof(OpenConnectionAsync)}");
         }
         
@@ -215,15 +226,13 @@ namespace Marvin.Migrations.PostgreSQL
             using (var connection = new NpgsqlConnection(connectionString))
             {
                 await connection.OpenAsync();
-                await InternalExecuteScriptAsync(script);
+                await InternalExecuteScriptAsync(connection, script);
             }
         }
 
-        private async Task InternalExecuteScriptAsync(string script)
+        private async Task InternalExecuteScriptAsync(NpgsqlConnection connection, string script)
         {
-            AssertConnection();
-            
-            var command = _connection.CreateCommand();
+            var command = connection.CreateCommand();
             command.CommandText = script;
             await command
                 .ExecuteNonQueryAsync();
@@ -232,7 +241,7 @@ namespace Marvin.Migrations.PostgreSQL
         /// <inheritdoc />
         public async Task CreateHistoryTableIfNotExistsAsync()
         {
-            AssertConnection();
+            AssertConnection(_connection);
             
             var script = $"CREATE TABLE IF NOT EXISTS public.\"{MigrationHistoryTableName}\" "
                          + @"( 
@@ -243,7 +252,7 @@ namespace Marvin.Migrations.PostgreSQL
                         ); ";
             try
             {
-                await ExecuteScriptAsync(script);
+                await InternalExecuteScriptAsync(_connection, script);
             }
             catch (PostgresException e)
                 when (e.SqlState.StartsWith("08")
@@ -264,6 +273,10 @@ namespace Marvin.Migrations.PostgreSQL
             {
                 throw new MigrationException(MigrationError.ConnectionError, $"Can not connect to DB {DbName}", e);
             }
+            catch (InvalidOperationException)
+            {
+                throw;
+            }
             catch (Exception e)
             {
                 throw new MigrationException(MigrationError.CreatingHistoryTable, $"Can not create database {DbName}",
@@ -275,6 +288,7 @@ namespace Marvin.Migrations.PostgreSQL
         public async Task<bool> CheckIfTableExistsAsync(string tableName)
         {
             if (String.IsNullOrWhiteSpace(tableName)) throw new ArgumentNullException(nameof(tableName));
+            AssertConnection(_connection);
             
             try
             {
@@ -283,8 +297,8 @@ namespace Marvin.Migrations.PostgreSQL
                                                        FROM   information_schema.tables "
                                                + $" WHERE  table_schema = '{GetSchemeNameFromConnectionString()}'"
                                                + $" AND    table_name = '{tableName}');";
-                var result = await ExecuteScalarScriptAsync(checkTableExistenceQuery);
-                return result != null && (Int32) result == 1;
+                var result = await InternalExecuteScalarScriptAsync(_connection, checkTableExistenceQuery);
+                return result != null && (result is int i && i == 1 || result is bool b && b);
             }
             catch (PostgresException e)
                 when (e.SqlState.StartsWith("08")
@@ -300,6 +314,10 @@ namespace Marvin.Migrations.PostgreSQL
             {
                 throw new MigrationException(MigrationError.AuthorizationError,
                     $"Invalid authorization specification for {DbName}", e);
+            }
+            catch (InvalidOperationException)
+            {
+                throw;
             }
             catch (Exception e)
             {
@@ -360,7 +378,7 @@ namespace Marvin.Migrations.PostgreSQL
         /// <inheritdoc />
         public async Task UpdateCurrentDbVersionAsync(DbVersion version)
         {
-            AssertConnection();
+            AssertConnection(_connection);
             
             var script = $"DELETE FROM public.\"{_options.MigrationHistoryTableName}\"; "
                          + $"INSERT INTO public.\"{_options.MigrationHistoryTableName}\"("
@@ -369,7 +387,7 @@ namespace Marvin.Migrations.PostgreSQL
 
             try
             {
-                await ExecuteScriptAsync(script);
+                await InternalExecuteScriptAsync(_connection, script);
             }
             catch (PostgresException e)
                 when (e.SqlState.StartsWith("08")
@@ -389,6 +407,10 @@ namespace Marvin.Migrations.PostgreSQL
             catch (NpgsqlException e)
             {
                 throw new MigrationException(MigrationError.ConnectionError, $"Can not connect to DB {DbName}", e);
+            }
+            catch (InvalidOperationException)
+            {
+                throw;
             }
             catch (Exception e)
             {
@@ -399,11 +421,11 @@ namespace Marvin.Migrations.PostgreSQL
         /// <inheritdoc />
         public async Task ExecuteScriptAsync(string script)
         {
-            AssertConnection();
+            AssertConnection(_connection);
             
             try
             {
-                await ExecuteScriptAsync(script);
+                await InternalExecuteScriptAsync(_connection, script);
             }
             catch (PostgresException e)
                 when (e.SqlState.StartsWith("08")
@@ -424,6 +446,10 @@ namespace Marvin.Migrations.PostgreSQL
             {
                 throw new MigrationException(MigrationError.ConnectionError, $"Can not connect to DB {DbName}", e);
             }
+            catch (InvalidOperationException)
+            {
+                throw;
+            }
             catch (Exception e)
             {
                 throw new MigrationException(MigrationError.MigratingError, "Can not execute script", e);
@@ -433,7 +459,7 @@ namespace Marvin.Migrations.PostgreSQL
         /// <inheritdoc />
         public async Task<object> ExecuteScalarScriptAsync(string script)
         {
-            AssertConnection();
+            AssertConnection(_connection);
             
             try
             {
@@ -459,6 +485,10 @@ namespace Marvin.Migrations.PostgreSQL
             {
                 throw new MigrationException(MigrationError.ConnectionError, $"Can not connect to DB {DbName}", e);
             }
+            catch (InvalidOperationException)
+            {
+                throw;
+            }
             catch (Exception e)
             {
                 throw new MigrationException(MigrationError.MigratingError, "Can not execute script", e);
@@ -475,8 +505,6 @@ namespace Marvin.Migrations.PostgreSQL
         /// <inheritdoc />
         public async Task ExecuteScriptWithoutInitialCatalogAsync(string script)
         {
-            AssertConnection();
-            
             try
             {
                 await ExecuteScriptAsync(_connectionStringWithoutInitialCatalog, script);
@@ -500,6 +528,10 @@ namespace Marvin.Migrations.PostgreSQL
             {
                 throw new MigrationException(MigrationError.ConnectionError, $"Can not connect to DB {DbName}", e);
             }
+            catch (InvalidOperationException)
+            {
+                throw;
+            }
             catch (Exception e)
             {
                 throw new MigrationException(MigrationError.MigratingError, "Can not execute script", e);
@@ -509,15 +541,13 @@ namespace Marvin.Migrations.PostgreSQL
         /// <inheritdoc />
         public async Task<object> ExecuteScalarScriptWithoutInitialCatalogAsync(string script)
         {
-            AssertConnection();
-            
             try
             {
                 using (var connection = new NpgsqlConnection(_connectionStringWithoutInitialCatalog))
                 {
                     await connection.OpenAsync();
                     var result = await InternalExecuteScalarScriptAsync(connection, script);
-                    
+
                     return result;
                 }
             }
@@ -539,6 +569,10 @@ namespace Marvin.Migrations.PostgreSQL
             catch (NpgsqlException e)
             {
                 throw new MigrationException(MigrationError.ConnectionError, $"Can not connect to DB {DbName}", e);
+            }
+            catch (InvalidOperationException)
+            {
+                throw;
             }
             catch (Exception e)
             {