+++ src/Marvin.Migrations.PostgreSQL/Changelog.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.2.6] - 2018-11-07
+
+### Changed
+
+- Assert connections only for script with initial catalogue.  
+
+## [0.2.5] - 2018-11-06
+
+### Fixed
+
+- Fixed `StackOverflowException` in  `PostgreDbProvider`
+
 ## [0.2.4] - 2018-11-06
 
 ### Added